+++ Unity/Packages/YooAsset/Runtime/AssetSystem/Loader/RawBundleFileLoader.cs
@@ -54,6 +54,11 @@ namespace YooAsset
 					_steps = ESteps.CheckFile;
 					FileLoadPath = MainBundleInfo.Bundle.CachedDataFilePath;
 				}
+				else if (MainBundleInfo.LoadMode == BundleInfo.ELoadMode.LoadFromDelivery)
+				{
+					_steps = ESteps.CheckFile;
+					FileLoadPath = MainBundleInfo.DeliveryFilePath;
+				}
 				else
 				{
 					throw new System.NotImplementedException(MainBundleInfo.LoadMode.ToString());
@@ -64,7 +69,8 @@ namespace YooAsset
 			if (_steps == ESteps.Download)
 			{
 				int failedTryAgain = Impl.DownloadFailedTryAgain;
-				_downloader = DownloadSystem.BeginDownload(MainBundleInfo, failedTryAgain);
+				_downloader = DownloadSystem.CreateDownload(MainBundleInfo, failedTryAgain);
+				_downloader.SendRequest();
 				_steps = ESteps.CheckDownload;
 			}
 
@@ -93,7 +99,8 @@ namespace YooAsset
 			{
 				int failedTryAgain = Impl.DownloadFailedTryAgain;
 				var bundleInfo = ManifestTools.ConvertToUnpackInfo(MainBundleInfo.Bundle);
-				_unpacker = DownloadSystem.BeginDownload(bundleInfo, failedTryAgain);
+				_unpacker = DownloadSystem.CreateDownload(bundleInfo, failedTryAgain);
+				_unpacker.SendRequest();
 				_steps = ESteps.CheckUnpack;
 			}