+++ Unity/Packages/YooAsset/Runtime/AssetSystem/Provider/ProviderBase.cs
@@ -123,7 +123,7 @@ namespace YooAsset
 		/// <summary>
 		/// 销毁资源对象
 		/// </summary>
-		public virtual void Destroy()
+		public void Destroy()
 		{
 			IsDestroyed = true;