+++ Unity/Packages/YooAsset/Runtime/Services/IBuildinQueryServices.cs
@@ -1,10 +1,10 @@
 ﻿
 namespace YooAsset
 {
-	public interface IQueryServices
+	public interface IBuildinQueryServices
 	{
 		/// <summary>
-		/// 查询内置资源
+		/// 查询应用程序里的内置资源是否存在
 		/// </summary>
 		bool QueryStreamingAssets(string packageName, string fileName);
 	}